Moved TypeAdapters and AbstractJSONDatabaseHandler in database/json package
