+++ src/main/java/world/bentobox/bentobox/database/json/adapters/WorldAdapter.java
@@ -1,18 +1,25 @@
-package world.bentobox.bentobox.database.mysql.adapters;
+package world.bentobox.bentobox.database.json.adapters;
 
 import java.io.IOException;
 
-import org.bukkit.potion.PotionEffectType;
+import org.bukkit.World;
+import org.bukkit.plugin.Plugin;
 
 import com.google.gson.TypeAdapter;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonToken;
 import com.google.gson.stream.JsonWriter;
 
-public class PotionEffectTypeAdapter extends TypeAdapter<PotionEffectType> {
+public class WorldAdapter extends TypeAdapter<World> {
+
+    private Plugin plugin;
+
+    public WorldAdapter(Plugin plugin) {
+        this.plugin = plugin;
+    }
 
     @Override
-    public void write(JsonWriter out, PotionEffectType value) throws IOException {
+    public void write(JsonWriter out, World value) throws IOException {
         if (value == null) {
             out.nullValue();
             return;
@@ -22,11 +29,11 @@ public class PotionEffectTypeAdapter extends TypeAdapter<PotionEffectType> {
     }
     
     @Override
-    public PotionEffectType read(JsonReader reader) throws IOException {
+    public World read(JsonReader reader) throws IOException {
         if (reader.peek() == JsonToken.NULL) {
             reader.nextNull();
             return null;
         }
-        return PotionEffectType.getByName(reader.nextString());
+        return plugin.getServer().getWorld(reader.nextString());
     }
 }
